src/backend/requirements.txt
feedparser==6.0.11
requests==2.32.3
anthropic==0.42.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
numpy>=1.22,<2.0
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.11.0
